Replace race_condition_fallback with a default_value_lambda and a value_updater_lambda
Motivation: for a herd of requests to different entries in a table, the race condition fallback strategy would cause the table to become unresponsive, as many long-running reads are scheduled.
To avoid such a situation, schedule these reads in a background process. To support that, support passing in a lambda to update the value in the cache. Default value of the lambda more or less preserves the old behavior.