Make `concurrent-ruby` a development (not runtime) dependency
It was added as a runtime dependency in
https://github.com/academia-edu/cached_counts/pull/17. It **is** a
runtime dependency, but only a transitive one (through Rails). The only
reason we need to specify it is that on Rails versions previous to Rails
7.1 or https://github.com/rails/rails/tree/7-0-stable (see
https://github.com/rails/rails/pull/54264 for the relevant fix), you'd
get an error with `concurrent-ruby` 1.3.5.

Once we no longer test against Rails 7.0, we'll probably be able to
remove this altogether.